agent-learning/ai-examples/requirements.txt
# LangChain + LangGraph Learning Requirements
# Install with: pip install -r requirements.txt

# Core LangChain packages
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.10

# LangGraph for workflows
langgraph==0.0.20

# Additional utilities
openai>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# Optional: for async support
aiohttp>=3.8.0
asyncio-throttle>=1.0.0

# Optional: for enhanced functionality
tiktoken>=0.5.0
numpy>=1.24.0
pandas>=2.0.0

# Development and testing
pytest>=7.0.0
pytest-asyncio>=0.21.0